Updated for current Nim 0.10.3
I am not sure for which version of the compiler your code was. I could not compile it with either 0.10.2 nor 0.10.3 so I changed it like this to compile it with 0.10.3 (2015-02-08). As I am new to Nim I would like to get a comment from you about what happened here? It looks like that there is some default constructor generated but that the way it works was changed since your code. Or do I miss something else?